[Decomp][Player] Cohesion split: extract reputation methods to PlayerReputation.cpp
