+++ src/game/Object/PlayerReputation.cpp
@@ -1,0 +1,250 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+
+
+#include "Player.h"
+#include "Language.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "SpellMgr.h"
+#include "World.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "UpdateMask.h"
+#include "SkillDiscovery.h"
+#include "QuestDef.h"
+#include "GossipDef.h"
+#include "UpdateData.h"
+#include "Channel.h"
+#include "ChannelMgr.h"
+#include "MapManager.h"
+#include "MapPersistentStateMgr.h"
+#include "InstanceData.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "CreatureAI.h"
+#include "Formulas.h"
+#include "Group.h"
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "Pet.h"
+#include "Util.h"
+#include "Transports.h"
+#include "Weather.h"
+#include "BattleGround/BattleGround.h"
+#include "BattleGround/BattleGroundMgr.h"
+#include "BattleGround/BattleGroundAV.h"
+#include "OutdoorPvP/OutdoorPvP.h"
+#include "ArenaTeam.h"
+#include "Chat.h"
+#include "revision_data.h"
+#include "Database/DatabaseImpl.h"
+#include "Spell.h"
+#include "ScriptMgr.h"
+#include "SocialMgr.h"
+#include "Mail.h"
+#include "SpellAuras.h"
+#include "DBCStores.h"
+#include "SQLStorages.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+/**
+ * @brief Gets the player's current reputation rank with a faction.
+ *
+ * @param faction The faction identifier to query.
+ * @return The current reputation rank.
+ */
+ReputationRank Player::GetReputationRank(uint32 faction) const
+{
+    FactionEntry const* factionEntry = sFactionStore.LookupEntry(faction);
+    return GetReputationMgr().GetRank(factionEntry);
+}
+
+// Calculate total reputation percent player gain with quest/creature level
+int32 Player::CalculateReputationGain(ReputationSource source, int32 rep, int32 faction, uint32 creatureOrQuestLevel, bool noAuraBonus)
+{
+    float percent = 100.0f;
+
+    float repMod = noAuraBonus ? 0.0f : (float)GetTotalAuraModifier(SPELL_AURA_MOD_REPUTATION_GAIN);
+
+    // faction specific auras only seem to apply to kills
+    if (source == REPUTATION_SOURCE_KILL)
+    {
+        repMod += GetTotalAuraModifierByMiscValue(SPELL_AURA_MOD_FACTION_REPUTATION_GAIN, faction);
+    }
+
+    percent += rep > 0 ? repMod : -repMod;
+
+    float rate;
+    switch (source)
+    {
+        case REPUTATION_SOURCE_KILL:
+            rate = sWorld.getConfig(CONFIG_FLOAT_RATE_REPUTATION_LOWLEVEL_KILL);
+            break;
+        case REPUTATION_SOURCE_QUEST:
+            rate = sWorld.getConfig(CONFIG_FLOAT_RATE_REPUTATION_LOWLEVEL_QUEST);
+            break;
+        case REPUTATION_SOURCE_SPELL:
+        default:
+            rate = 1.0f;
+            break;
+    }
+
+    if (rate != 1.0f && creatureOrQuestLevel <= MaNGOS::XP::GetGrayLevel(getLevel()))
+    {
+        percent *= rate;
+    }
+
+    if (percent <= 0.0f)
+    {
+        return 0;
+    }
+
+    // Multiply result with the faction specific rate
+    if (const RepRewardRate* repData = sObjectMgr.GetRepRewardRate(faction))
+    {
+        float repRate = 0.0f;
+        switch (source)
+        {
+            case REPUTATION_SOURCE_KILL:
+                repRate = repData->creature_rate;
+                break;
+            case REPUTATION_SOURCE_QUEST:
+                repRate = repData->quest_rate;
+                break;
+            case REPUTATION_SOURCE_SPELL:
+                repRate = repData->spell_rate;
+                break;
+        }
+
+        // for custom, a rate of 0.0 will totally disable reputation gain for this faction/type
+        if (repRate <= 0.0f)
+        {
+            return 0;
+        }
+
+        percent *= repRate;
+    }
+
+    return int32(sWorld.getConfig(CONFIG_FLOAT_RATE_REPUTATION_GAIN) * rep * percent / 100.0f);
+}
+
+// Calculates how many reputation points player gains in victim's enemy factions
+void Player::RewardReputation(Unit* pVictim, float rate)
+{
+    if (!pVictim || pVictim->GetTypeId() == TYPEID_PLAYER)
+    {
+        return;
+    }
+
+    if (((Creature*)pVictim)->IsReputationGainDisabled())
+    {
+        return;
+    }
+
+    // used current difficulty creature entry instead normal version (GetEntry())
+    ReputationOnKillEntry const* Rep = sObjectMgr.GetReputationOnKillEntry(((Creature*)pVictim)->GetCreatureInfo()->Entry);
+
+    if (!Rep)
+    {
+        return;
+    }
+
+    if (Rep->repfaction1 && (!Rep->team_dependent || GetTeam() == ALLIANCE))
+    {
+        int32 donerep1 = CalculateReputationGain(REPUTATION_SOURCE_KILL, Rep->repvalue1, Rep->repfaction1, pVictim->getLevel());
+        donerep1 = int32(donerep1 * rate);
+        FactionEntry const* factionEntry1 = sFactionStore.LookupEntry(Rep->repfaction1);
+        uint32 current_reputation_rank1 = GetReputationMgr().GetRank(factionEntry1);
+        if (factionEntry1 && current_reputation_rank1 <= Rep->reputation_max_cap1)
+        {
+            GetReputationMgr().ModifyReputation(factionEntry1, donerep1);
+        }
+
+        // Wiki: Team factions value divided by 2
+        if (factionEntry1 && Rep->is_teamaward1)
+        {
+            FactionEntry const* team1_factionEntry = sFactionStore.LookupEntry(factionEntry1->team);
+            if (team1_factionEntry)
+            {
+                GetReputationMgr().ModifyReputation(team1_factionEntry, donerep1 / 2);
+            }
+        }
+    }
+
+    if (Rep->repfaction2 && (!Rep->team_dependent || GetTeam() == HORDE))
+    {
+        int32 donerep2 = CalculateReputationGain(REPUTATION_SOURCE_KILL, Rep->repvalue2, Rep->repfaction2, pVictim->getLevel());
+        donerep2 = int32(donerep2 * rate);
+        FactionEntry const* factionEntry2 = sFactionStore.LookupEntry(Rep->repfaction2);
+        uint32 current_reputation_rank2 = GetReputationMgr().GetRank(factionEntry2);
+        if (factionEntry2 && current_reputation_rank2 <= Rep->reputation_max_cap2)
+        {
+            GetReputationMgr().ModifyReputation(factionEntry2, donerep2);
+        }
+
+        // Wiki: Team factions value divided by 2
+        if (factionEntry2 && Rep->is_teamaward2)
+        {
+            FactionEntry const* team2_factionEntry = sFactionStore.LookupEntry(factionEntry2->team);
+            if (team2_factionEntry)
+            {
+                GetReputationMgr().ModifyReputation(team2_factionEntry, donerep2 / 2);
+            }
+        }
+    }
+}
+
+// Calculate how many reputation points player gain with the quest
+void Player::RewardReputation(Quest const* pQuest)
+{
+    // quest reputation reward/loss
+    for (int i = 0; i < QUEST_REPUTATIONS_COUNT; ++i)
+    {
+        if (!pQuest->RewRepFaction[i])
+        {
+            continue;
+        }
+
+        if (pQuest->RewRepValue[i])
+        {
+            int32 rep = CalculateReputationGain(REPUTATION_SOURCE_QUEST,  pQuest->RewRepValue[i], pQuest->RewRepFaction[i], GetQuestLevelForPlayer(pQuest));
+
+            if (FactionEntry const* factionEntry = sFactionStore.LookupEntry(pQuest->RewRepFaction[i]))
+            {
+                GetReputationMgr().ModifyReputation(factionEntry, rep);
+            }
+        }
+    }
+
+    // TODO: implement reputation spillover
+}
